feat: move all banner related files to banner component dir
