+++ src/components/KonfluxBanner/BannerContent.tsx
@@ -3,22 +3,23 @@ import { Flex } from '@patternfly/react-core';
 import { BellIcon } from '@patternfly/react-icons/dist/esm/icons/bell-icon';
 import { ExclamationTriangleIcon } from '@patternfly/react-icons/dist/esm/icons/exclamation-triangle-icon';
 import { InfoCircleIcon } from '@patternfly/react-icons/dist/esm/icons/info-circle-icon';
+import { BannerConfig, BannerType, BannerVariant } from '~/components/KonfluxBanner/banner-type';
 import { SyncMarkdownView } from '~/shared/components/markdown-view/MarkdownView';
-import { BannerConfig, BannerType } from '~/types/banner';
 
 type BannerContentProps = Pick<BannerConfig, 'type' | 'summary'>;
 
-const typeToIcon = (type: BannerType) => {
-  switch (type) {
-    case 'info':
-      return <InfoCircleIcon data-test="info-icon" />;
-    case 'warning':
-      return <BellIcon data-test="warning-icon" />;
-    case 'danger':
-      return <ExclamationTriangleIcon data-test="danger-icon" />;
-    default:
-      return <InfoCircleIcon data-test="info-icon" />;
-  }
+// Map BannerType → BannerVariant
+export const bannerTypeToVariant: Record<BannerType, BannerVariant> = {
+  info: BannerVariant.BLUE,
+  warning: BannerVariant.GOLD,
+  danger: BannerVariant.RED,
+};
+
+// Map BannerVariant → Icon component
+const bannerVariantToIcon: Record<BannerVariant, JSX.Element> = {
+  [BannerVariant.BLUE]: <InfoCircleIcon data-test="info-icon" />,
+  [BannerVariant.GOLD]: <BellIcon data-test="warning-icon" />,
+  [BannerVariant.RED]: <ExclamationTriangleIcon data-test="danger-icon" />,
 };
 
 export const BannerContent: React.FC<BannerContentProps> = React.memo(({ type, summary }) => (
@@ -27,9 +28,8 @@ export const BannerContent: React.FC<BannerContentProps> = React.memo(({ type, s
     justifyContent={{ default: 'justifyContentCenter' }}
     flexWrap={{ default: 'nowrap' }}
     spaceItems={{ default: 'spaceItemsSm' }}
-    style={{ width: '100%' }}
   >
-    {typeToIcon(type)}
+    {bannerVariantToIcon[bannerTypeToVariant[type]]}
     <SyncMarkdownView content={summary} inline />
   </Flex>
 ));
