--- src/consts/banner.ts
@@ -1,2 +1,0 @@
-export const BANNER_NAMEPSACE = 'konflux-info';
-export const BANNER_CONTENT_FILE = 'konflux-banner-configmap';
