+++ src/consts/constants.ts
@@ -3,3 +3,4 @@ export const REPO_PUSH = 'repo_push';
 export const PLACEHOLDER = '<PLACEHOLDER>';
 export const NO_RESULTS = 'no results';
 export const TYPEAHEAD_THRESHOLD = 10;
+export const KONFLUX_INFO_NAMESPACE = 'konflux-info';
